requirements.txt
torch
torchvision
numpy
pillow
matplotlib
seaborn
scikit-learn
python-chess
cairosvg
opencv-python
tqdm
pyautogui
stockfish
